adjust component composition of NavBar
I wanted to adjust the composition so that NavBar is the parent of the
SearchBar and Filters components. This will be more semantically
correct, as the h1 element and the search form can now all be rendered
inside a header element, where the styling can be cohesive and the
components that actually are the navigation elements of the app
(SearchBar and Filters), now are rendered in NavBar.